requirements.txt
numpy==1.24.1
opencv_python==4.7.0.68
Pillow==9.4.0
scipy==1.10.0
torchvision==0.14.1
Wand==0.6.10
